requirements.txt
numpy<2.0.0
scipy==1.14.0
matplotlib==3.8.2
pyyaml==6.0.1
shapely==2.0.5
cvxpy==1.5.3
sympy==1.13.0
reprep-z7==7.2
zuper-commons-z7==7.2
PyGeometry-z7==7.2
pycontracts3==7.2
do-mpc==4.6.5
pandas
typing-extensions
frozendict==2.4.4
xtermcolor
pytz
aiofiles
future
webcolors
decorator
pulp>=2.7.0
qdldl==0.1.7.post5
osqp==0.6.3
casadi==3.6.5
toolz
cytoolz
networkx==3.3
commonroad-drivability-checker==2024.1
cachetools

triangle @ git+https://github.com/ClearCalcs/triangle.git@master
